Fix "back" behavior
Added "onStack" flag to correctly deal with unwinding the stack.
When the flag is true, the client skips looking for the response in the
cache and simply displays the response on the back of the stack.
"onStack" is set as true when a back command is given and reset false
after the resource resolution/cache check step.